Refactor to be branchless.
Based on: https://blog.roblox.com/2017/04/voxel-terrain-storage/

Also get rid of the support for non-u32 types; consumers all have different
needs when it comes to translating. In future, we may make this generic
overall, e.g. to store floating point data.

The only heavyweight thing left is the hashtable, but that's easy to fix later.