Let's use random values as object ids instead
As the comment says, we need unique across the network and time, and we don't
need to be as fast as everything else. Uuids are the industry-standard way of
doing this, but we don't need all of that: just 1 bit for the ephemeral flag
and one caarved out value for allowing it to play nice with Option.  We can just
feed that from a cryptographic random generator, and give the mint a few ways
to generate them deterministically for testing.