Position, our first component. Fixes to the derive macro to make this happen.
The derive now takes strings which must contain expressions like derivative:
Rust won't let us include the expressions directly.  What we do here is
generally use a constant, then put the name of that constant in the string:
since other code is going to need to refer to them by name, we'd be doing that
anyway.