introduce asset_store
Right now, this is an abstracted interface over files, but later it will do a lot more. The point of it
is that we will eventually want to do lots of fun things, most notably
encryption (as much as I hate doing that) and incremental updates,
plus eager loading of assets stored in more efficient formats
like zip files.